Integrated interlocking_bahndsl into handler_controller
